move everything into one folder
